worker/requirements.txt
fastapi
uvicorn
pypdf
python-dotenv
psycopg2-binary
requests
groq
python-multipart